LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-to-visible-areas-margin-y-axis-expected.txt


FAIL Scroll margin should be considered when calculating snap area visibilty while snapping on the y-axis assert_equals: expected 700 but got 500
